Attempt to make CUDA future with allocator compile
